code review feedback
move getSchoolLocationsCalculatedBenefits to utils
